+++ core.osgi/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/osgi/commands/TestRepairTableStatusCommand.java
@@ -186,7 +186,7 @@ public class TestRepairTableStatusCommand
         TableReference tableReference = createTableRef("ks1.tbl1");
         RepairStateSnapshot state = mock(RepairStateSnapshot.class);
         when(state.getVnodeRepairStates()).thenReturn(vnodeRepairStates);
-        return new ScheduledRepairJobView(UUID.randomUUID(), tableReference, null, state, RepairJobView.Status.IN_QUEUE, 0);
+        return new ScheduledRepairJobView(UUID.randomUUID(), tableReference, null, state, RepairJobView.Status.IN_QUEUE, 0, 0);
     }
 
     private RepairTableStatusCommand mockCommand(String tableRef)