+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/metrics/TableMetricHolder.java
@@ -35,6 +35,7 @@ public class TableMetricHolder implements Closeable
     static final String REPAIR_TIMING_FAILED = "RepairFailedTime";
     static final String LAST_REPAIRED_AT = "LastRepairedAt";
     static final String REPAIR_STATE = "RepairState";
+    static final String REMAINING_REPAIR_TIME = "RemainingRepairTime";
 
     private final MetricRegistry myMetricRegistry;
     private final NodeMetricHolder myNodeMetricHolder;
@@ -43,6 +44,7 @@ public class TableMetricHolder implements Closeable
 
     private final AtomicReference<RangeRepairState> myRepairState = new AtomicReference<>();
     private final AtomicReference<Long> myLastRepairedAt = new AtomicReference<>(0L);
+    private final AtomicReference<Long> myRemainingRepairTime = new AtomicReference<>(0L);;
 
     public TableMetricHolder(TableReference tableReference, MetricRegistry metricRegistry, NodeMetricHolder nodeMetricHolder)
     {
@@ -69,6 +71,7 @@ public class TableMetricHolder implements Closeable
             }
         });
         myMetricRegistry.register(metricName(LAST_REPAIRED_AT), lastRepairedAtGauge());
+        myMetricRegistry.register(metricName(REMAINING_REPAIR_TIME), remainingRepairTimeGauge());
         timer(REPAIR_TIMING_SUCCESS);
         timer(REPAIR_TIMING_FAILED);
     }
@@ -100,6 +103,11 @@ public class TableMetricHolder implements Closeable
         myLastRepairedAt.set(lastRepairedAt);
     }
 
+    public void remainingRepairTime(long remainingRepairTime)
+    {
+        myRemainingRepairTime.set(remainingRepairTime);
+    }
+
     public void repairTiming(long timeTaken, TimeUnit timeUnit, boolean successful)
     {
         if (successful)
@@ -121,6 +129,7 @@ public class TableMetricHolder implements Closeable
         myMetricRegistry.remove(metricName(REPAIR_TIMING_FAILED));
         myMetricRegistry.remove(metricName(LAST_REPAIRED_AT));
         myMetricRegistry.remove(metricName(REPAIR_STATE));
+        myMetricRegistry.remove(metricName(REMAINING_REPAIR_TIME));
     }
 
     private String metricName(String name)
@@ -137,6 +146,11 @@ public class TableMetricHolder implements Closeable
     private Gauge<Long> lastRepairedAtGauge()
     {
         return myLastRepairedAt::get;
+    }
+
+    private Gauge<Long> remainingRepairTimeGauge()
+    {
+        return myRemainingRepairTime::get;
     }
 
     private static class RangeRepairState