+++ rest/src/test/java/com/ericsson/bss/cassandra/ecchronos/rest/TestRepairManagementRESTImpl.java
@@ -441,7 +441,7 @@ public class TestRepairManagementRESTImpl
         // Given
         RepairConfiguration repairConfig = TestUtils.createRepairConfiguration(11, 2.2, 33, 44);
         RepairJobView repairJobView = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0, 0);
         TableRepairConfig expectedResponse = new TableRepairConfig(repairJobView);
 
         when(myRepairScheduler.getCurrentRepairJobs()).thenReturn(Collections.singletonList(repairJobView));
@@ -458,11 +458,11 @@ public class TestRepairManagementRESTImpl
         // Given
         RepairConfiguration repairConfig = TestUtils.createRepairConfiguration(11, 2.2, 33, 44);
         RepairJobView repairJobView = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0, 0);
 
         RepairConfiguration repairConfig2 = TestUtils.createRepairConfiguration(22, 3.3, 44, 55);
         RepairJobView repairJobView2 = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks2", "tbl"), repairConfig2, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks2", "tbl"), repairConfig2, null, Status.IN_QUEUE, 0, 0);
 
         RepairJobView repairJobView3 = new OnDemandRepairJobView(UUID.randomUUID(),
                 myTableReferenceFactory.forTable("ks", "tbl"), RepairConfiguration.DEFAULT, Status.IN_QUEUE, 0,
@@ -499,7 +499,7 @@ public class TestRepairManagementRESTImpl
     {
         RepairConfiguration repairConfig = TestUtils.createRepairConfiguration(11, 2.2, 33, 44);
         RepairJobView repairJobView = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0, 0);
 
         when(myRepairScheduler.getCurrentRepairJobs()).thenReturn(Collections.singletonList(repairJobView));
 
@@ -514,7 +514,7 @@ public class TestRepairManagementRESTImpl
     {
         RepairConfiguration repairConfig = TestUtils.createRepairConfiguration(11, 2.2, 33, 44);
         RepairJobView repairJobView = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0, 0);
 
         TableRepairConfig expectedResponse = new TableRepairConfig(repairJobView);
 
@@ -531,11 +531,11 @@ public class TestRepairManagementRESTImpl
     {
         RepairConfiguration repairConfig = TestUtils.createRepairConfiguration(11, 2.2, 33, 44);
         RepairJobView repairJobView = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0, 0);
 
         RepairConfiguration repairConfig2 = TestUtils.createRepairConfiguration(22, 3.3, 44, 55);
         RepairJobView repairJobView2 = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl2"), repairConfig2, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl2"), repairConfig2, null, Status.IN_QUEUE, 0, 0);
 
         List<TableRepairConfig> expectedResponse = Arrays.asList(
                 new TableRepairConfig(repairJobView),
@@ -566,7 +566,7 @@ public class TestRepairManagementRESTImpl
     {
         RepairConfiguration repairConfig = TestUtils.createRepairConfiguration(11, 2.2, 33, 44);
         RepairJobView repairJobView = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0, 0);
 
         when(myRepairScheduler.getCurrentRepairJobs()).thenReturn(Collections.singletonList(repairJobView));
 
@@ -581,7 +581,7 @@ public class TestRepairManagementRESTImpl
     {
         RepairConfiguration repairConfig = TestUtils.createRepairConfiguration(11, 2.2, 33, 44);
         RepairJobView repairJobView = new ScheduledRepairJobView(UUID.randomUUID(),
-                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0);
+                myTableReferenceFactory.forTable("ks", "tbl"), repairConfig, null, Status.IN_QUEUE, 0, 0);
 
         List<TableRepairConfig> expectedResponse = Collections.singletonList(new TableRepairConfig(repairJobView));